OVS-3249 moved deploy saio script as well as deploy esx/kvm machines to separate file
